refactor(paths): extract launchd paths into lib/launchd-paths.ts
